requirements.txt
ffmpeg-python
google-cloud-speech
google-cloud-storage
